+++ molo/core/models.py
@@ -55,7 +55,6 @@ from molo.core.utils import (
 )
 
 from django.db.models.signals import pre_delete
-from wagtail.contrib.forms.models import AbstractEmailForm, AbstractFormField
 from django_enumfield import enum
 
 
@@ -1149,11 +1148,6 @@ class Main(CommentedPageMixin, MoloPage):
                     generate_slug(self.title), )))
             self.add_child(instance=footer_index)
             footer_index.save_revision().publish()
-            form_index = FormIndexPage(
-                title='Forms', slug=('forms-%s' % (
-                    generate_slug(self.title), )))
-            self.add_child(instance=form_index)
-            form_index.save_revision().publish()
             tag_index = TagIndexPage(
                 title='Tags', slug=('tags-%s' % (
                     generate_slug(self.title), )))
@@ -2015,58 +2009,3 @@ FooterPage.promote_panels = [
     MultiFieldPanel(
         Page.promote_panels,
         "Common page configuration", "collapsible collapsed")]
-
-
-class FormIndexPage(MoloPage, PreventDeleteMixin):
-    parent_page_types = []
-    subpage_types = ['FormPage']
-
-    def copy(self, *args, **kwargs):
-        site = kwargs['to'].get_site()
-        main = site.root_page
-        FormIndexPage.objects.child_of(main).delete()
-        super(FormIndexPage, self).copy(*args, **kwargs)
-
-    def get_site(self):
-        try:
-            return self.get_ancestors().filter(
-                depth=2).first().sites_rooted_here.get(
-                    site_name__icontains='main')
-        except Exception:
-            return self.get_ancestors().filter(
-                depth=2).first().sites_rooted_here.all().first() or None
-
-
-class FormField(AbstractFormField):
-    page = ParentalKey(
-        'FormPage', on_delete=models.CASCADE, related_name='form_fields')
-
-
-class FormPage(TranslatablePageMixinNotRoutable, AbstractEmailForm):
-    parent_page_types = ['FormIndexPage']
-    subpage_types = []
-    language = models.ForeignKey(
-        'core.SiteLanguage', blank=True, null=True,
-        on_delete=models.SET_NULL)
-    translated_pages = models.ManyToManyField("self", blank=True)
-    intro = models.TextField(blank=True)
-    body = StreamField([
-        ('paragraph', blocks.RichTextBlock()),
-    ], null=True, blank=True)
-    thank_you_text = models.TextField(blank=True)
-    content_panels = AbstractEmailForm.content_panels + [
-        FieldPanel('intro', classname="full"),
-        StreamFieldPanel('body'),
-        InlinePanel('form_fields', label="Form fields"),
-        FieldPanel('thank_you_text', classname="full"),
-        MultiFieldPanel([
-            FieldRowPanel([
-                FieldPanel('from_address', classname="col6"),
-                FieldPanel('to_address', classname="col6"),
-            ]),
-            FieldPanel('subject'),
-        ], "Email"),
-    ]
-
-    def serve(self, request, *args, **kwargs):
-        return super(AbstractEmailForm, self).serve(request, *args, **kwargs)