+++ molo/core/tests/test_template_tags.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 from django.test import TestCase, RequestFactory
 from molo.core.models import (
     Main, SiteLanguageRelation, Languages, BannerPage, ArticlePageTags,
-    FormPage, SiteSettings, ArticleOrderingChoices, ReactionQuestionChoice,
+    SiteSettings, ArticleOrderingChoices, ReactionQuestionChoice,
     ReactionQuestion, ReactionQuestionResponse, ReactionQuestionIndexPage)
 from molo.core.tests.base import MoloTestCaseMixin
 from molo.core.templatetags.core_tags import (
@@ -15,7 +15,6 @@ from molo.core.templatetags.core_tags import (
     load_child_articles_for_section, load_reaction_choice_submission_count,
     load_user_choice_reaction_question, get_tag_articles
 )
-from molo.core.templatetags.forms_tags import forms_list
 
 
 @pytest.mark.django_db
@@ -45,29 +44,6 @@ class TestModels(TestCase, MoloTestCaseMixin):
         self.factory = RequestFactory()
         self.request = self.factory.get('/')
         self.request.site = self.site
-
-    def create_form_page(
-            self, parent, title="Test Form",
-            slug="test-form", **kwargs):
-        form_page = FormPage(
-            title=title,
-            slug=slug,
-            intro='Introduction to Test Form ...',
-            thank_you_text='Thank you for taking the Test Form',
-            **kwargs
-        )
-        parent.add_child(instance=form_page)
-        form_page.save_revision().publish()
-        return form_page
-
-    def test_get_form_list_homepage(self):
-        context = {
-            'locale_code': 'en',
-            'request': self.request,
-            'forms': self.create_form_page(self.form_index)
-        }
-        context = forms_list(context)
-        self.assertEqual(len(context['forms']), 1)
 
     def test_load_user_choice_reaction_question(self):
         article = self.mk_articles(self.yourmind, 1)[0]